create SurveyFilesManager actor to handle generation of files
